+++ WolvenKit.App/Functionality/Services/IWatcherService.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using DynamicData;
 using WolvenKit.Models;
 using WolvenKit.MVVM.Model.ProjectManagement.Project;
+using WolvenKit.ViewModels.Editor.Basic;
 
 namespace WolvenManager.App.Services
 {
@@ -15,6 +16,8 @@ namespace WolvenManager.App.Services
         public bool IsSuspended { get; set; }
 
         //public IObservable<IChangeSet<FileModel, ulong>> Connect();
+        public IObservable<IChangeSet<FileViewModel, ulong>> Connect();
+
         public IObservableCache<FileModel, ulong> Files { get; }
 
         public Task RefreshAsync(EditorProject proj);
